requirements.txt
accelerate==0.30.1
diffusers==0.32.2
einops==0.8.1
face_alignment==1.4.1
facenet_pytorch==2.6.0
imageio==2.37.0
insightface==0.7.3
ipdb==0.13.13
kornia==0.8.0
loguru==0.7.3
mediapipe==0.10.21
moviepy==2.1.2
numpy==2.2.5
omegaconf==2.3.0
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.2.1
pyarrow==19.0.0
pytorch3d==0.7.8
PyYAML==6.0.2
PyYAML==6.0.2
safetensors==0.5.3
scipy==1.15.2
setuptools==60.2.0
setuptools==75.8.0
skimage==0.0
timm==0.6.13
torchfile==0.1.0
tqdm==4.67.1
transformers==4.49.0
xformers==0.0.28
yacs==0.1.8
